fix stack traces on linux
